Fin de la gestion de API de paiement

+++ database/migrations/2025_05_22_140302_create_comptes_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('comptes', function (Blueprint $table) {
             $table->uuid('id')->primary();
             $table->uuid('entreprise_id');
-            $table->integer('montant');
+            $table->decimal('montant', 12, 2);
             $table->timestamps();
         });
     }
